use any transform node as eye frame

+++ OpenGL/FoundationOfCG/Assignment4/Assignment4-RobotsAndPartPicking/Assignment4-RobotsAndPartPicking/RobotsAndPartPicking/scenegraph.h
@@ -49,7 +49,7 @@ public:
   void removeChild(shared_ptr<SgNode> child);
 
   int getNumChildren() const {
-    return children_.size();
+    return (int)children_.size();
   }
 
   shared_ptr<SgNode> getChild(int i) {
